Added FL to combined Positions popup
